+++ CHANGELOG.md
@@ -9,11 +9,13 @@
 ### ⚙️ Miscellaneous Tasks
 
 - Fmt
+
 ## [1.3.1] - 2026-04-06
 
 ### 🐛 Bug Fixes
 
 - Make sure files are executable
+
 ## [1.3.0] - 2026-04-06
 
 ### 🚀 Features
@@ -32,7 +34,8 @@
 ### ⚙️ Miscellaneous Tasks
 
 - Bump
-- *(release)* Super-release-windows-x64@1.1.0, super-release-linux-arm64@1.1.0, super-release-darwin-arm64@1.1.0, super-release@1.3.0, super-release-darwin-x64@1.1.0, super-release-linux-x64@1.1.0 [skip ci]
+- _(release)_ Super-release-windows-x64@1.1.0, super-release-linux-arm64@1.1.0, super-release-darwin-arm64@1.1.0, super-release@1.3.0, super-release-darwin-x64@1.1.0, super-release-linux-x64@1.1.0 [skip ci]
+
 ## [1.2.0] - 2026-04-04
 
 ### 🚀 Features
@@ -44,6 +47,7 @@
 - Configured toolchain
 - Adjusted ci/cd, added dependabot
 - Formatting, added more tests for monorepos
+
 ## [1.1.4] - 2026-04-04
 
 ### 🐛 Bug Fixes
@@ -53,6 +57,7 @@
 ### ⚙️ Miscellaneous Tasks
 
 - Added lockfile
+
 ## [1.1.3] - 2026-04-01
 
 ### 🐛 Bug Fixes
@@ -69,12 +74,14 @@
 - Run tests in parallel, fixed output, stream results
 - Add failure outputs
 - Enforce CI=false on gh action runners as well
+
 ## [1.1.2] - 2026-04-01
 
 ### 🐛 Bug Fixes
 
 - Some cleanup / refactoring
 - Fail on invalid / missing maintenance branch ranges
+
 ## [1.1.1] - 2026-04-01
 
 ### 🐛 Bug Fixes
@@ -84,52 +91,62 @@
 ### ⚙️ Miscellaneous Tasks
 
 - Added 2e2 comparison
+
 ## [1.1.0] - 2026-04-01
 
 ### 🚀 Features
 
 - Added more maintenance / channel tests and adjustments, added docs for different scenarios
 - Added version collision tests / handling
+
 ## [1.0.3] - 2026-04-01
 
 ### 🐛 Bug Fixes
 
 - Make sure to handle prerelease cutoff correctly
+
 ## [1.0.2] - 2026-04-01
 
 ### 🐛 Bug Fixes
 
 - Simplifications
+
 ## [1.0.1] - 2026-04-01
 
 ### 🐛 Bug Fixes
 
 - Make sure not to rewrite the `package.json` files and preserve formatting
+
 ## [1.0.0] - 2026-04-01
 
 ### 🚀 Features
 
 - [**breaking**] Added support for json and json config files, cleanup
+
 ## [0.15.0] - 2026-04-01
 
 ### 🚀 Features
 
 - Better npm precheck error handling, updated docs
+
 ## [0.14.0] - 2026-04-01
 
 ### 🚀 Features
 
 - Run publish checks in parallel
+
 ## [0.13.3] - 2026-04-01
 
 ### 🐛 Bug Fixes
 
 - Adjusted checks, more output
+
 ## [0.13.2] - 2026-04-01
 
 ### 🐛 Bug Fixes
 
 - Made output clearer
+
 ## [0.13.1] - 2026-04-01
 
 ### 🐛 Bug Fixes
@@ -139,36 +156,43 @@
 ### ⚙️ Miscellaneous Tasks
 
 - Added test for npm skip publish case
+
 ## [0.13.0] - 2026-04-01
 
 ### 🚀 Features
 
 - Test if the version is already published before publishing
+
 ## [0.12.0] - 2026-04-01
 
 ### 🚀 Features
 
 - Show multi progress for concurrent builds
+
 ## [0.11.0] - 2026-04-01
 
 ### 🚀 Features
 
 - Add package include filter for branches
+
 ## [0.10.0] - 2026-03-31
 
 ### 🚀 Features
 
 - Add support for musl
+
 ## [0.9.0] - 2026-03-31
 
 ### 🚀 Features
 
 - Simplified & optimized using `git log`, fixed config file handling
+
 ## [0.8.0] - 2026-03-31
 
 ### 🚀 Features
 
 - Review & performance improvements
+
 ## [0.7.0] - 2026-03-31
 
 ### 🚀 Features
@@ -178,16 +202,19 @@
 ### ⚙️ Miscellaneous Tasks
 
 - Format
+
 ## [0.6.0] - 2026-03-31
 
 ### 🚀 Features
 
 - Refactored command exec, make clearer in what folder the commands are executed
+
 ## [0.5.1] - 2026-03-31
 
 ### 🐛 Bug Fixes
 
 - Moved package update to the right location, added info about package includes/excludes
+
 ## [0.5.0] - 2026-03-31
 
 ### 🚀 Features
@@ -197,16 +224,19 @@
 ### ⚙️ Miscellaneous Tasks
 
 - Docs
+
 ## [0.4.0] - 2026-03-31
 
 ### 🚀 Features
 
 - Communicate version properly, adjusted docs, add `--show-next-version`
+
 ## [0.3.1] - 2026-03-31
 
 ### 🐛 Bug Fixes
 
 - Escape properly
+
 ## [0.3.0] - 2026-03-31
 
 ### 🚀 Features
@@ -216,12 +246,14 @@
 ### ⚙️ Miscellaneous Tasks
 
 - Make sure not to bump version for no-bump commits
+
 ## [0.2.1] - 2026-03-31
 
 ### ⚙️ Miscellaneous Tasks
 
 - Cleanup
 - Removed artifacts, cleanup
+
 ## [0.2.0] - 2026-03-30
 
 ### 🚀 Features